requirements.txt
absl-py==0.5.0
astor==0.7.1
cycler==0.10.0
decorator==4.3.2
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
numpy==1.15.2
protobuf==3.6.1
PyYAML==5.1.2
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
